// src/lib/richtext-renderer/richtext-types.generated.ts
export interface TiptapNodeAttributes {
  paragraph: { textAlign?: 'left' | 'center' | 'right' | 'justify' | null; };
  doc: Record<string, never>;
  text: Record<string, never>;
  blockquote: Record<string, never>;
  heading: { textAlign?: 'left' | 'center' | 'right' | 'justify' | null; level?: 1 | 2 | 3 | 4 | 5 | 6; };
  bullet_list: Record<string, never>;
  ordered_list: { start?: any; type?: any; order?: number; };
  list_item: Record<string, never>;
  code_block: { class?: string | null; };
  hard_break: Record<string, never>;
  horizontal_rule: Record<string, never>;
  image: { src?: string; alt?: string; title?: string | null; width?: number | string | null; height?: number | string | null; id?: string | null; source?: any; copyright?: any; meta_data?: any; };
  emoji: { name?: string | null; emoji?: string | null; fallbackImage?: string | null; };
  table: Record<string, never>;
  tableRow: Record<string, never>;
  tableCell: { colspan?: number; rowspan?: number; colwidth?: number[] | null; backgroundColor?: string | null; };
  tableHeader: { colspan?: number; rowspan?: number; colwidth?: number[] | null; };
  blok: { id?: string | null; body?: SbBlokData[]; };
  details: Record<string, never>;
  detailsContent: Record<string, never>;
  detailsSummary: Record<string, never>;
}

/** Attribute types for all Tiptap mark extensions */
export interface TiptapMarkAttributes {
  link: { href?: string; target?: '_self' | '_blank' | '_parent' | '_top' | null; rel?: string | null; class?: string | null; title?: string | null; linktype?: 'url' | 'story' | 'asset' | 'email'; };
  textStyle: { backgroundColor?: string | null; color?: string | null; fontFamily?: any; fontSize?: any; lineHeight?: any; };
  bold: Record<string, never>;
  italic: Record<string, never>;
  strike: Record<string, never>;
  underline: Record<string, never>;
  code: Record<string, never>;
  superscript: Record<string, never>;
  subscript: Record<string, never>;
  highlight: { color?: string | null; };
  anchor: { id?: string | null; };
  styled: { class?: string | null; };
  reporter: Record<string, never>;
}

export type TiptapNodeName = keyof TiptapNodeAttributes;
export type TiptapMarkName = keyof TiptapMarkAttributes;